requirements.txt
Django==1.5.4
django-model-utils==1.5.0
djangorestframework==2.3.8
git+https://github.com/mikedrawback/djangorestframework-docjson.git